Introduce boost mode
This change introduces new configuration options for a "boost mode".

The poll interval should be short enough to process messages in time, and long enough to minimize resource consumption for the database. Boost mode is a solution which automatically handles spikes by processing messages in quick succession. When no new messages appear for a specified time, boost mode is disabled again.

Through this change, `pollInterval` now also accepts fractions of seconds, for example "0.5" for half a second.